Replace 'long' by 'int' to avoid bug on Linux

Commit migrated from https://github.com/dotnet/coreclr/commit/54913c983114fb451c8c55b3c51290b5512a4e59

+++ src/coreclr/src/jit/morph.cpp
@@ -11076,7 +11076,7 @@ T GetSignedMagicNumberForDivide(T denom, int *shift /*out*/)
     int iters = 0;
 
     absDenom = abs(denom);
-    t = two_nminus1 + ((unsigned long)denom >> 31);
+    t = two_nminus1 + ((unsigned int)denom >> 31);
     absNc = t - 1 - (t % absDenom);     // absolute value of nc
     p = bits_minus_1;                   // initialize p
     q1 = two_nminus1 / absNc;           // initialize q1 = 2^p / abs(nc)
